+++ django-backend/fecfiler/memo_text/serializers.py
@@ -2,10 +2,8 @@ from .models import MemoText
 from django.db import transaction
 from fecfiler.validation import serializers
 from rest_framework.serializers import UUIDField, ModelSerializer
-from fecfiler.committee_accounts.serializers import (
-    CommitteeOwnedSerializer,
-    ReportCommitteeValidationMixin,
-)
+from fecfiler.committee_accounts.serializers import CommitteeOwnedSerializer
+from fecfiler.reports.serializers import ReportCommitteeValidationMixin
 import structlog
 
 logger = structlog.get_logger(__name__)
@@ -88,3 +86,20 @@ class LinkedMemoTextSerializerMixin(ModelSerializer):
             memo_object = MemoText.objects.get(id=memo_text_id)
             memo_object.delete()
             validated_data["memo_text_id"] = None
+
+    def validate_memo_text_committee_ownership(self, data):
+        committee_account = data.get("committee_account")
+        id = data.get("memo_text_id")
+        if id:
+            exists = MemoText.objects.filter(
+                id=id, committee_account=committee_account
+            ).exists()
+            if not exists:
+                raise serializers.ValidationError(
+                    {
+                        "memo_text_id": (
+                            "Invalid memo_text_id or memo text does not belong "
+                            "to this committee account."
+                        )
+                    }
+                )
